test: make it a shared datastore test
